move: Auth 관련 어노테이션 코드 presentation -> infrastructure 패키지로 이동
