+++ payment/src/main/java/com/nowayback/payment/infrastructure/config/WebConfig.java
@@ -1,7 +1,7 @@
 package com.nowayback.payment.infrastructure.config;
 
-import com.nowayback.payment.presentation.auth.role.RoleCheckInterceptor;
-import com.nowayback.payment.presentation.auth.user.AuthUserArgumentResolver;
+import com.nowayback.payment.infrastructure.auth.role.RoleCheckInterceptor;
+import com.nowayback.payment.infrastructure.auth.user.AuthUserArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
